+++ src/tenable/createAssetExportCache.ts
@@ -1,18 +1,8 @@
 import { AssetExportCache } from '.';
 import TenableClient from './TenableClient';
-import {
-  AssetExport,
-  ExportAssetsOptions,
-  ExportStatus,
-} from '@jupiterone/tenable-client-nodejs';
+import { AssetExport } from '@jupiterone/tenable-client-nodejs';
 
-import {
-  IntegrationError,
-  IntegrationLogger,
-} from '@jupiterone/integration-sdk-core';
-import { sleep } from '@lifeomic/attempt';
-import { addMinutes, isAfter } from 'date-fns';
-import pMap from 'p-map';
+import { IntegrationLogger } from '@jupiterone/integration-sdk-core';
 
 export async function createAssetExportCache(
   logger: IntegrationLogger,
@@ -34,35 +24,9 @@ export async function createAssetExportCache(
 }
 
 async function getAssetsUsingExport(client: TenableClient) {
-  const options: ExportAssetsOptions = { chunk_size: 100 };
-  const { export_uuid: exportUuid } = await client.exportAssets(options);
-  let { status, chunks_available: chunksAvailable } =
-    await client.fetchAssetsExportStatus(exportUuid);
-
-  const timeLimit = addMinutes(Date.now(), 30);
-  while ([ExportStatus.Processing, ExportStatus.Queued].includes(status)) {
-    ({ status, chunks_available: chunksAvailable } =
-      await client.fetchAssetsExportStatus(exportUuid));
-
-    if (isAfter(Date.now(), timeLimit)) {
-      await client.cancelAssetExport(exportUuid);
-      throw new IntegrationError({
-        code: 'TenableClientApiError',
-        message: `Asset export ${exportUuid} failed to finish processing in time limit`,
-      });
-    }
-    await sleep(60_000); // Sleep 1 minute between status checks.
-  }
-
-  const chunkResponses = await pMap(
-    chunksAvailable,
-    async (chunkId) => await client.fetchAssetsExportChunk(exportUuid, chunkId),
-    { concurrency: 3 },
-  );
-
-  const assetExports = chunkResponses.reduce((prev, cur) => {
-    return prev.concat(cur);
-  }, []);
-
+  const assetExports: AssetExport[] = [];
+  await client.iterateAssets((a) => {
+    assetExports.push(a);
+  });
   return assetExports;
 }