+++ src/constants.ts
@@ -2,11 +2,13 @@ import { RelationshipClass } from '@jupiterone/integration-sdk-core';
 
 export const SetDataKeys = {
   USERS: 'data-users',
+  ASSET_MAP: 'data-asset-map',
 };
 
 export const StepIds = {
   ACCOUNT: 'step-account',
   SCANS: 'step-scans',
+  ASSETS: 'step-assets',
   SCAN_DETAILS: 'step-scan-details',
   USER_SCAN_RELATIONSHIPS: 'step-user-scan-relationships',
   USERS: 'step-users',
@@ -80,6 +82,12 @@ export const relationships = {
     sourceType: entities.USER._type,
     _class: RelationshipClass.OWNS,
     targetType: entities.SCAN._type,
+  },
+  SCAN_SCANS_ASSET: {
+    _type: 'tenable_scan_scans_asset',
+    sourceType: entities.SCAN._type,
+    _class: RelationshipClass.SCANS,
+    targetType: 'aws_instance,azure_vm,google_compute_instance,tenable_asset',
   },
   ACCOUNT_HAS_CONTAINER: {
     _type: 'tenable_account_has_container',