+++ CHANGELOG.md
@@ -10,6 +10,18 @@ and this project adheres to
 
 ### Added
 
+- Added support for ingesting the following **new** resources:
+
+  | Resources | Entity `_type`  | Entity `_class` |
+  | --------- | --------------- | --------------- |
+  | Asset     | `tenable_asset` | `Host`          |
+
+- Added support for ingesting the following **new** relationships:
+
+  | Source         | \_class | Target          |
+  | -------------- | ------- | --------------- |
+  | `tenable_scan` | `SCANS` | `tenable_asset` |
+
 - Added `TenableClient.iteraetAssets()` method. Made other asset export
   endpoints private.