+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to
 
 ### Added
 
+- Added support for ingesting the following **new** relationships:
+
+  | Source         | \_class   | Target                                                        |
+  | -------------- | --------- | ------------------------------------------------------------- |
+  | `tenable_scan` | **SCANS** | `aws_instance,azure_vm,google_compute_instance,tenable_asset` |
+
 - Added `TenableClient.iterateAssets()` method. Made other asset export
   endpoints private.